AOC 2023 day 12 - part 2 with memoization

+++ src/main/java/be/rouget/puzzles/adventofcode/year2023/day12/RunningDamagedSpringGroup.java
@@ -63,10 +63,10 @@ public record RunningDamagedSpringGroup(boolean inGroup, Integer groupIndex, int
             return damagedSprings.isEmpty();
         }
 
-        if (groupIndex.intValue() + 1 != damagedSprings.size()) {
+        if (groupIndex + 1 != damagedSprings.size()) {
             return false;
         }
 
-        return groupLength == damagedSprings.get(groupIndex.intValue()).intValue();
+        return groupLength == damagedSprings.get(groupIndex);
     }
 }
